automl-agent/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "automl-agent"
version = "1.0.0"
description = "Revolutionary multi-agent AutoML platform with specialized collaborative intelligence"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "AutoML Agent Team", email = "team@automl-agent.ai"}
]
keywords = ["automl", "machine-learning", "agents", "ai", "crewai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers", 
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10", 
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.9"
dependencies = [
    # Core Dependencies
    "pandas>=2.0.0",
    "numpy>=1.21.0", 
    "scikit-learn>=1.3.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.12.0",
    "scipy>=1.9.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "requests>=2.28.0",
    
    # Machine Learning Algorithms
    "xgboost>=2.0.0",
    "lightgbm>=4.0.0",
    "catboost>=1.2.0",
    "tensorflow>=2.13.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    
    # Agent Orchestration
    "crewai>=0.28.0",
    "langchain>=0.1.0",
    "langchain-community>=0.0.20",
    "langchain-openai>=0.0.8",
    "openai>=1.12.0",
    
    # Hyperparameter Optimization
    "optuna>=3.5.0",
    "hyperopt>=0.2.7",
    "ray[tune]>=2.8.0",
    "scikit-optimize>=0.9.0",
    "bayesian-optimization>=1.4.0",
    
    # Experiment Tracking
    "mlflow>=2.10.0",
    "wandb>=0.16.0",
    "langsmith>=0.1.0",
    
    # Natural Language Processing
    "nltk>=3.8.0",
    "spacy>=3.7.0",
    "transformers>=4.36.0",
    "datasets>=2.16.0",
    "sentence-transformers>=2.2.0",
    
    # Computer Vision
    "opencv-python>=4.8.0",
    "pillow>=10.0.0",
    "albumentations>=1.3.0",
    
    # Time Series
    "statsmodels>=0.14.0",
    "prophet>=1.1.0",
    "pmdarima>=2.0.0",
    "sktime>=0.24.0",
    
    # Recommendation Systems
    "scikit-surprise>=1.1.0",
    "implicit>=0.7.0",
    
    # Data Preprocessing
    "feature-engine>=1.6.0",
    "category-encoders>=2.6.0",
    "imbalanced-learn>=0.11.0",
    "missingno>=0.5.0",
    "fancyimpute>=0.7.0",
    
    # Clustering & Dimensionality Reduction
    "umap-learn>=0.5.0",
    "hdbscan>=0.8.0",
    
    # Model Interpretation
    "shap>=0.44.0",
    "lime>=0.2.0",
    "eli5>=0.13.0",
    
    # API Framework
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "aiofiles>=23.2.1",
    "websockets>=12.0",
    
    # Web Interface
    "streamlit>=1.30.0",
    "gradio>=4.0.0",
    
    # Data Validation
    "pydantic>=2.5.0",
    "great-expectations>=0.18.0",
    
    # Cloud Integration
    "boto3>=1.34.0",
    "sagemaker>=2.190.0",
    "s3fs>=2023.12.0",
    
    # Visualization
    "plotly>=5.17.0",
    "bokeh>=3.3.0",
    "wordcloud>=1.9.0",
    
    # Data Formats
    "pyarrow>=14.0.0",
    "openpyxl>=3.1.0",
    "xlsxwriter>=3.1.0",
    "h5py>=3.10.0",
    
    # Utilities
    "tqdm>=4.66.0",
    "joblib>=1.3.0",
    "dask>=2023.12.0",
    
    # Statistical Analysis
    "pingouin>=0.5.0",
    "lifelines>=0.27.0",
    
    # Database & Storage
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    
    # Monitoring & Logging
    "prometheus-client>=0.19.0",
    "loguru>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]

jupyter = [
    "jupyter>=1.0.0",
    "notebook>=7.0.0",
    "ipykernel>=6.25.0",
]

extra-automl = [
    "autogluon>=1.0.0",
    "auto-sklearn>=0.15.0",
    "flaml>=2.1.0", 
    "pycaret>=3.2.0",
]

all = [
    "automl-agent[dev,jupyter,extra-automl]"
]

[project.urls]
Homepage = "https://github.com/automl-agent/automl-agent"
Documentation = "https://automl-agent.readthedocs.io"
Repository = "https://github.com/automl-agent/automl-agent.git"
"Bug Tracker" = "https://github.com/automl-agent/automl-agent/issues"

[project.scripts]
automl-agent = "src.cli:main"
automl-api = "src.api.main:main"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = [
    "--cov=src",
    "--cov-report=html",
    "--cov-report=term-missing",
]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = true
ignore_missing_imports = true
